refactor: dispatch core.Flow & devkit.Flow
