+++ src/promptflow/promptflow/_sdk/_submitter/test_submitter.py
@@ -13,7 +13,7 @@ from colorama import Fore, init
 from promptflow._internal import ConnectionManager
 from promptflow._proxy import ProxyFactory
 from promptflow._sdk._constants import PROMPT_FLOW_DIR_NAME
-from promptflow._sdk._utils import dump_flow_result, parse_variant
+from promptflow._sdk._utils import parse_variant
 from promptflow._sdk.entities._flow import Flow, FlowContext
 from promptflow._sdk.operations._local_storage_operations import LoggerOperations
 from promptflow._utils.context_utils import _change_working_dir
@@ -28,10 +28,11 @@ from ..._constants import LINE_NUMBER_KEY, FlowLanguage
 from ..._core._errors import NotSupported
 from ..._utils.async_utils import async_run_allowing_running_loop
 from ..._utils.dataclass_serializer import convert_eager_flow_output_to_dict
+from ..._utils.flow_utils import dump_flow_result
 from ..._utils.logger_utils import get_cli_sdk_logger
 from ...batch import APIBasedExecutorProxy, CSharpExecutorProxy
 from .._configuration import Configuration
-from ..entities._eager_flow import FlexFlow
+from ..entities._flow import FlexFlow
 from .utils import (
     SubmitterHelper,
     print_chat_output,
