+++ src/promptflow/promptflow/core/_flow_context_resolver.py
@@ -125,15 +125,21 @@ class FlowContextResolver:
             with open(flow_file, "w") as fp:
                 dump_yaml(self.flow_dag, fp)
             resolved_flow._path = flow_file.absolute().as_posix()
+
+            executable_flow = resolved_flow._init_executable()
             if is_async_call:
                 return AsyncFlowInvoker(
-                    flow=resolved_flow,
+                    flow=executable_flow,
                     connections=connections,
                     streaming=flow_context.streaming,
+                    flow_path=resolved_flow.path,
+                    working_dir=resolved_flow.code,
                 )
             else:
                 return FlowInvoker(
-                    flow=resolved_flow,
+                    flow=executable_flow,
                     connections=connections,
                     streaming=flow_context.streaming,
+                    flow_path=resolved_flow.path,
+                    working_dir=resolved_flow.code,
                 )
