+++ src/promptflow/tests/sdk_cli_test/unittests/test_flow.py
@@ -7,8 +7,7 @@ import pytest
 from marshmallow import ValidationError
 
 from promptflow import load_flow
-from promptflow._sdk.entities._eager_flow import FlexFlow
-from promptflow._sdk.entities._flow import Flow
+from promptflow._sdk.entities._flow import FlexFlow, Flow
 
 FLOWS_DIR = Path("./tests/test_configs/flows")
 EAGER_FLOWS_DIR = Path("./tests/test_configs/eager_flows")
